refactor(agents): rewire modules after installer/auth/readiness moves
Compile-green milestone for the regroup: agents/mod.rs drops the moved
top-level modules; readiness::resolver consumers move to
readiness::service (sessions, model_registry, probe, launch_options);
seed and npm reach the lock and managed_npm helpers through installer;
auth_config declares login_terminal and its un-#[path]-hacked test
files; installer/tests.rs gets explicit imports.

622 lib tests pass; boundary, max-lines, and old-path checks green.
