+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/managed_npm.rs
@@ -109,7 +109,7 @@ fn recorded_source_spec(managed_dir: &Path) -> Option<String> {
     }
 }
 
-pub(super) fn managed_npm_install_issue(package: &str, managed_dir: &Path) -> Option<String> {
+pub(crate) fn managed_npm_install_issue(package: &str, managed_dir: &Path) -> Option<String> {
     if is_npm_non_registry_spec(package) {
         if non_registry_install_matches(package, managed_dir) {
             return None;
