+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/npm.rs
@@ -2,7 +2,7 @@ use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 
 use super::{InstallError, InstalledArtifactResult};
-use crate::domains::agents::managed_npm::{
+use super::managed_npm::{
     apply_npm_version_override, installed_npm_package_version, managed_npm_install_issue,
     npm_package_version, write_managed_npm_source_marker,
 };
