+++ anyharness/crates/anyharness-lib/src/domains/agents/mod.rs
@@ -1,10 +1,7 @@
 pub mod auth_config;
 pub mod catalog;
 pub mod credentials;
-pub mod install_lock;
 pub mod installer;
-pub mod login_terminal;
-mod managed_npm;
 pub mod model;
 pub mod model_registry;
 pub mod portability;
