+++ anyharness/crates/anyharness-lib/src/domains/agents/readiness/mod.rs
@@ -3,5 +3,5 @@ mod compatibility;
 pub mod launch_options;
 mod overrides;
 pub(crate) mod paths;
-pub mod resolver;
+pub mod service;
 mod status;
