+++ anyharness/crates/anyharness-lib/src/domains/agents/runtime.rs
@@ -7,7 +7,7 @@ use super::model::*;
 use super::readiness::paths::{
     artifact_root, managed_registry_binary_for_names, managed_registry_npm_binary_for_names,
 };
-use super::readiness::resolver::resolve_agent;
+use super::readiness::service::resolve_agent;
 use super::reconcile::execution::{AgentReconcileJobSnapshot, AgentReconcileService};
 use super::registry::built_in_registry;
 use super::seed::AgentSeedStore;
